+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Set Ansible config hash behaviour to merge
 * Add config-examples-* make targets
 * Add CloudWatch logging for Dispatchers (Author and Publish), Orchestrator and Chaos Monkey
+* Replace Stack Builder AMI IDs config generation Python script with Ansible module
 
 ### 2.4.0
 * Explicit installation of libtool, autoconf, and automake for native package compilation